pyproject.toml
[tool.poetry]
name = "Facial-recognition"
version = "0.1.0"
description = ""
authors = [
  "Ivan Garcia <ivanjesusgarcia1@gmail.com>",
  "Jose Contreras <contrerasgedler@gmail.com>",
]


[[tool.poetry.packages]]
include = "facial_recognition"
from = "src"

[tool.poetry.dependencies]
python = "^3.9"
pytest = ">=8.3.5,<9.0.0"
pylint = ">=3.3.7,<4.0.0"
black = "^25.1.0"
torch = "^2.2.2+cu121"
Pillow = "^10.2.0"
opencv-python = "^4.11.0"
numpy = "1.26.4"
facenet_pytorch = "2.5.0"
flask = "^3.1.1"
gunicorn = "^23.0.0"
psutil = "^7.0.0"
flask-cors = "^6.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
registers = "facial_recognition.registers:main"
access = "facial_recognition.access:main"
flask = "facial_recognition.app:app.run(debug=True)"
